net: shell: Add virtual command
Move "virtual" command to a separate file.
