+++ subsys/net/lib/shell/CMakeLists.txt
@@ -28,6 +28,7 @@ zephyr_library_sources(stats.c)
 zephyr_library_sources(suspend.c)
 zephyr_library_sources(tcp.c)
 zephyr_library_sources(udp.c)
+zephyr_library_sources(virtual.c)
 
 zephyr_library_include_directories(
   ${ZEPHYR_BASE}/kernel/include
